requirements.txt
spacy>=3.0.0
nltk>=3.6.0
pdfplumber>=0.5.28
python-docx>=0.8.11
regex>=2021.4.4
flask>=2.0.0
werkzeug>=2.0.0
reportlab>=3.6.12 